Migrate away from custom dependency submission (#56)
GitHub previously did not natively support dependency submission for Gradle projects and we had to use https://github.com/mikepenz/gradle-dependency-submission in order to get Dependabot to know about our dependencies and flag security issues. It looks like that recently changed and now GitHub's Gradle-Build action supports this natively. So this code migrates to that approach.