Remove keycloakAdapterCore, add BouncyCastle and Jetbrains (for @VisibleForTesting), bump to Java 17, prepare for Keycloak 25 (#143)
Jumping straight to Keycloak 25 broke the build. It seems the "keycloakAdapterCore" package was removed.

Simply removing that dependency from this repo showed that we were dependent on it only for transitive dependencies - on Guava and BouncyCastle. I added BouncyCastle as a new dependency, and I replaced Guava with Jetbrains, just for the "VisibleForTesting" annotation.

As well, Keycloak 25 is on Java 17, so I bumped that up too. After everything, our tests pass so it looks like this should be a clean change.