Switch from Serverless to AWS CDK (#232)
Instead of using the Serverless framework for managing cloud resources and deployments, this project now uses the AWS CDK.

A few things were somewhat clunky and harder to do in the CDK than in Serverless, namely setting up the JWT Authorizer as well as the API route in APIGateway. We opted for using the CDK's HttpJwtAuthorizer class instead of the seemingly more relevant HttpUserPoolAuthorizer because the Cognito User Pool is not managed inside of this application. In the future, that may change, and/or I believe we can reference/import resources simply with their ARN, but for now, this approach is good enough and makes sense to me.

As well, the code goes through some extra work in order to create a custom Log Group. This was done only for the fact that it is the way Serverless used to handle Log Groups for Lambdas, and while it is unclear if there is benefit to it, it seems pretty common, standard, and easy, so that is also done here.

Note that in making this IaC migration and using the Lambda's same programming language in the same module, it is now possible to share string values between the infrastructure and the Lambda code through constants, and so the entire need for the Lambda's environment variables went away. This will not the be same for every project and every use-case (environment variables still have other good uses), but here, it was a welcome simplification.